+++ crates/azure_devops_types/src/lib.rs
@@ -18,7 +18,7 @@ mod azure_devops_service_endpoint_kind;
 mod azure_devops_service_endpoint_name;
 mod azure_devops_team;
 mod azure_devops_team_member;
-mod azure_devops_user_entitlement;
+mod azure_devops_user_license_entitlement;
 mod azure_devops_user_id;
 mod azure_devops_work_item_query;
 mod azure_devops_work_items;
@@ -54,7 +54,7 @@ pub mod prelude {
     pub use crate::azure_devops_service_endpoint_name::*;
     pub use crate::azure_devops_team::*;
     pub use crate::azure_devops_team_member::*;
-    pub use crate::azure_devops_user_entitlement::*;
+    pub use crate::azure_devops_user_license_entitlement::*;
     pub use crate::azure_devops_user_id::*;
     pub use crate::azure_devops_work_item_query::*;
     pub use crate::azure_devops_work_items::*;